+++ xtask/Cargo.toml
@@ -3,6 +3,9 @@ name = "xtask"
 version = "0.1.0"
 edition = "2018"
 
+[[bin]]
+name = "xtask"
+
 [dependencies]
 anyhow = "1.0.51"
